catkin_ws/src/spot_fsm_control/requirements.txt
bosdyn-api==3.2.0
bosdyn-choreography-client==3.2.0
bosdyn-choreography-protos==3.2.0
bosdyn-client==3.2.0
bosdyn-core==3.2.0
bosdyn-mission==3.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
grpcio==1.50.0
idna==3.4
numpy==1.23.4
Pillow==9.3.0
protobuf==4.21.9
PyJWT==2.6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
wrapt==1.14.1
python-dotenv
python-statemachine
rospy
omegaconf
scikit-learn
opencv-python
openvr
pygame